bpf plugins: fix plugins not stopping on context cancel
When the client event table changes, Velociraptor cancels the context
and expects plugins to shut themselves down. Then it starts them again
if they are still needed in the new table. The bpf based plugins were
not stopping on cancel, but Velociraptor was starting new goroutines for
them, and the old goroutines continued to run forever leaking resources.
This commit changes the bpf plugins so they exit when the context is
cancelled.