ensure bpf plugins shutdown fully before new start
When the client event table changes, velociraptor cancels the
context and the plugins catch this and start exiting. Unfortunately
velociraptor is not waiting for the plugins to shutdown fully before
starting them again. The bpf plugins use defer statements to close
the bpf modules, and these must run fully before the plugins are
started again, or there will be bpf module load failures. This commit
ensures a new instance of a bpf plugin cannot be started before the
previous one is fully done.