audit: only send events to interested subscribers
Events were being broadcast to all subscribers and it was up to them to
filter out irrelevant events in their WHERE clauses. With many audit
queries running and a high event rate, there is a performance overhead
due to extra copying over channels and goroutine scheduling.

Our artifacts either call the audit plugin with one key or no keys.
If called with one key then it only wants events tagged with that key.
If called with no keys then it only wants events that are untagged.